backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Pydantic (settings support)
pydantic==2.7.4
pydantic-settings==2.3.4

# Authentication
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# Environment variables
python-dotenv==1.0.1

# CORS
starlette==0.37.2

# Optional (recommended)
alembic==1.13.2